Control roles

+++ src/main/java/sistemas/tallerpoo/visual/RegistroPaciente.java
@@ -24,6 +24,8 @@ import sistemas.tallerpoo.clasesLogicas.Triage;
 import sistemas.tallerpoo.datos.AdmisionDatos;
 import javax.swing.JLabel;
 import javax.swing.JTextField;
+import sistemas.tallerpoo.clasesLogicas.ControlRoles;
+import sistemas.tallerpoo.clasesLogicas.Rol;
 
 /**
  *
@@ -500,6 +502,11 @@ public class RegistroPaciente extends javax.swing.JFrame {
     }// </editor-fold>//GEN-END:initComponents
 
     private void btnRegistrarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnRegistrarActionPerformed
+        if(!ControlRoles.usuarioTieneRol(new Rol("Medico", 0)) && !ControlRoles.usuarioTieneRol(new Rol("Admision de Pacientes", 0))){
+            JOptionPane.showMessageDialog(null, "Usted no tiene acceso a esta función");
+            return;
+        }
+        
         avisoDNI.setVisible(false);
         avisoNOMBRE.setVisible(false);
         avisoAPELLIDO.setVisible(false);
@@ -536,7 +543,7 @@ public class RegistroPaciente extends javax.swing.JFrame {
             }
             return;
         }
-
+      
         if ((Integer.parseInt(txtDni.getText()) <= 0)){
             JOptionPane.showMessageDialog(null,"Ingrese un DNI válido", "Error",JOptionPane.ERROR_MESSAGE);
             return;
@@ -545,6 +552,7 @@ public class RegistroPaciente extends javax.swing.JFrame {
               JOptionPane.showMessageDialog(null, "Correo electrónico incorrecto", "Error", JOptionPane.ERROR_MESSAGE);
             return;
         }
+      
         Admision a = new Admision();
         Paciente p = datos.capturar(txtDni, txtNombre, txtApellido, txtFechaNacimiento, txtDomicilio, txtTelefono, txtCelular, cbEstadoCivil, txtCorreo, txtContacto);
         if (datos.agregarPaciente(p)){
@@ -594,6 +602,11 @@ public class RegistroPaciente extends javax.swing.JFrame {
     }//GEN-LAST:event_btnMostrarTodoActionPerformed
 
     private void btnEditarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnEditarActionPerformed
+        if(!ControlRoles.usuarioTieneRol(new Rol("Medico", 0)) && !ControlRoles.usuarioTieneRol(new Rol("Admision de Pacientes", 0))){
+            JOptionPane.showMessageDialog(null, "Usted no tiene acceso a esta función");
+            return;
+        }
+        
         Paciente pac;
         int fila = jtPacientes.getSelectedRow();
         pac = lista.get(fila);
@@ -601,6 +614,11 @@ public class RegistroPaciente extends javax.swing.JFrame {
     }//GEN-LAST:event_btnEditarActionPerformed
 
     private void btnEliminarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnEliminarActionPerformed
+        if(!ControlRoles.usuarioTieneRol(new Rol("Medico", 0)) && !ControlRoles.usuarioTieneRol(new Rol("Admision de Pacientes", 0))){
+            JOptionPane.showMessageDialog(null, "Usted no tiene acceso a esta función");
+            return;
+        }
+        
         int fila = jtPacientes.getSelectedRow();
         String s = jtPacientes.getModel().getValueAt(fila, 0).toString();
         datos.eliminarPaciente(Integer.parseInt(s));
@@ -613,6 +631,11 @@ public class RegistroPaciente extends javax.swing.JFrame {
     }//GEN-LAST:event_txtBuscarDniKeyReleased
 
     private void btnConfirmarEdicionActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnConfirmarEdicionActionPerformed
+        if(!ControlRoles.usuarioTieneRol(new Rol("Medico", 0)) && !ControlRoles.usuarioTieneRol(new Rol("Admision de Pacientes", 0))){
+            JOptionPane.showMessageDialog(null, "Usted no tiene acceso a esta función");
+            return;
+        }
+        
         Paciente p = new Paciente();
         p=datos.capturar(txtDni, txtNombre, txtApellido, txtFechaNacimiento, txtDomicilio, txtDni, txtCelular, cbEstadoCivil, txtCorreo, txtContacto);
 
@@ -660,6 +683,11 @@ public class RegistroPaciente extends javax.swing.JFrame {
     }//GEN-LAST:event_txtFechaNacimientoFocusGained
 
     private void btnAdmisionActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnAdmisionActionPerformed
+        if(!ControlRoles.usuarioTieneRol(new Rol("Medico", 0)) && !ControlRoles.usuarioTieneRol(new Rol("Admision de Pacientes", 0))){
+            JOptionPane.showMessageDialog(null, "Usted no tiene acceso a esta función");
+            return;
+        }
+        
         int fila = jtPacientes.getSelectedRow();
         if(fila == -1){
             JOptionPane.showMessageDialog(null, "Debe seleccionar un paciente", "Error", JOptionPane.ERROR_MESSAGE);
